requirements.txt
boto3==1.26.141
botocore==1.29.141
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16